Update resvg to 0.34 for improved performance (#65)
* update resvg to 0.34.0

* Update baseline images

* Update rust licenses

* Refactor spec tests to write failed images to directory

* upload test failures

* Use dssim to check image similarity

+++ vl-convert-rs/src/text.rs
@@ -216,7 +216,7 @@ fn extract_text_width(svg: &String) -> Result<f64, AnyError> {
             if let Some(ref bbox) = path.text_bbox {
                 let width = bbox.right() - bbox.left();
                 let _height = bbox.bottom() - bbox.top();
-                return Ok(width);
+                return Ok(width as f64);
             }
         }
     }
